pyproject.toml
[tool.poetry]
name = "gameyamlspiderandgenerator"
version = "1.4.4"
description = "A useful tool for generating Furrygameindex yaml files"
authors = ["kaesinol"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langcodes = "^3.3.0"
language-data = "^1.1"
html2text = "^2020.1.16"
"ruamel.yaml" = "^0.17.21"
"ruamel.yaml.string" = "^0.1.0"
urllib3 = "^1.26.14"
beautifulsoup4 = "^4.11.1"
pillow = "^9.4.0"
requests = "^2.28.2"
ruamel-base = "^1.0.0"
loguru = "^0.6.0"
pyyaml = "^6.0"
pysocks = "^1.7.1"
epicstore-api = "^0.1.6"
jsonschema = "^4.17.3"
openai = "^0.27.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
